Add content adapter for Astro collections
