--- content/technologies/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "@rawkodeacademy/content-technologies",
-  "version": "0.1.0",
-  "private": true,
-  "type": "module",
-  "description": "Rawkode Academy technologies content and schema",
-  "main": "./src/index.ts",
-  "module": "./src/index.ts",
-  "types": "./src/index.ts",
-  "dependencies": {
-    "zod": "^3.24.3"
-  },
-  "exports": {
-    ".": "./src/index.ts",
-    "./package.json": "./package.json"
-  },
-  "files": [
-    "data/**/*",
-    "src/**/*",
-    "**/*.mdx",
-    "**/*.svg",
-    "**/*.ico"
-  ]
-}
