+++ projects/rawkode.academy/website/package.json
@@ -52,8 +52,7 @@
                 "@pothos/plugin-drizzle": "^0.13.0",
                 "@pothos/plugin-federation": "^4.3.2",
                 "@rawkodeacademy/content": "workspace:*",
-                "@rawkodeacademy/content-technologies": "workspace:*",
-		"@tailwindcss/forms": "^0.5",
+                "@tailwindcss/forms": "^0.5",
 		"@tailwindcss/vite": "^4.1.12",
 		"@trigger.dev/astro": "^3.3.12",
 		"@trigger.dev/react": "^3.3.12",
